+++ NEWS.md
@@ -1,3 +1,8 @@
+# dv.explorer.parameter 0.1.9-9000
+
+* lineplot:
+    * Added a function that ensures the indicated ref line variables in the input data table are of "labelled" class when a label is assigned to them.
+
 # dv.explorer.parameter 0.1.9
 
 * boxplot, lineplot, scatter plot, scatter plot matrix, correlation heatmap, waterfall plus heatmap :